refactor: FooterとCallbackコンポーネントをts化 #122
