+++ src/components/Footer.tsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 const Footer = () => {
   return (
     <>
